refactor: Insert MessageEditor and his components inside folder
